all feature   'origin/master'

+++ app/src/main/java/com/example/greendefend/domin/model/weather/List.kt
@@ -11,7 +11,7 @@ data class List (
   @SerializedName("clouds"     ) var clouds     : Clouds?            = Clouds(),
   @SerializedName("wind"       ) var wind       : Wind?              = Wind(),
   @SerializedName("visibility" ) var visibility : Int?               = null,
-  @SerializedName("pop"        ) var pop        : Int?               = null,
+  @SerializedName("pop"        ) var pop        : Float?               = null,
   @SerializedName("sys"        ) var sys        : Sys?               = Sys(),
   @SerializedName("dt_txt"     ) var dtTxt      : String?            = null
